part 2 | exercise 2.15

+++ part2/phonebook/src/components/Person.jsx
@@ -1,24 +1,22 @@
 import services from '../services/numbers'
 
-const Person = ({name, number, id, setPersons, persons}) => {
-  const deleteBtn = (id) => {
-    console.log(id);
-    services
-    .deletePhone(id)
-    .then(data =>{
-      console.log(data);
-      setPersons(persons.filter(item => item.id !== data.id))
-    })
-    .catch(err => console.log(err))
+const Person = ({name, number, id, setPersons}) => {
+
+  const deleteNumber = (id, name) => {
+    if (window.confirm(`Do you really want to delete ${name} ?`)) {
+      services
+      .deletePhone(id)
+      .then(data=>{
+        setPersons(prev => prev.filter(item => item.id !== data.id))
+      })
+    }
   }
+
+
   return (
-    <>
-      <p>{name} {number}</p>
-      <button onClick={(event) => {
-        event.preventDefault()
-        deleteBtn(id)
-      }}>Delete</button>
-    </>
+    <div key={id}>
+    {name} {number}<button onClick={() => deleteNumber(id, name)}>delete</button>
+    </div>
   )
 }
